Introduce Tuple dialect for manipulating tuple values.
This commit introduces a dialect around the builtin tuple type for
manipulating tuple values similar to how the complex and tensor dialects
provide ops for dealing with values of those types. We will use this
dialect as a data type for iterators that return "several values". We
currently do that using LLVM structs, which maily limits the types we
can use as elements but also composes badly with other passes like
bufferization.