Replace in-place IR modification with functions of the rewriter.
This is a requirement for RewritePatterns, and the previous
implementation of applySignatureConversion, which is meant to be called
by patterns, did not respect that. Since the new form requires knowing
the locations of converted values before the corresponding values are
converted, I also added two convenience functions for mapping locations.